+++ test/CMakeLists.txt
@@ -5,7 +5,7 @@ find_package(Boost REQUIRED)
 file(GLOB test_src CONFIGURE_DEPENDS "*.hpp" "*.cpp")
 
 
-add_executable (Test ${test_src})
+add_executable (Test ${test_src} test_bucket_distributor.cpp)
 
 
 target_link_libraries(Test PUBLIC "-Wl,--disable-new-dtags")
